+++ L2J_Mobius_CT_2.6_HighFive/java/com/l2jmobius/gameserver/model/base/SubClass.java
@@ -30,7 +30,7 @@ public final class SubClass
 	
 	private PlayerClass _class;
 	private long _exp = ExperienceData.getInstance().getExpForLevel(Config.BASE_SUBCLASS_LEVEL);
-	private int _sp = 0;
+	private long _sp = 0;
 	private byte _level = Config.BASE_SUBCLASS_LEVEL;
 	private int _classIndex = 1;
 	
@@ -71,7 +71,7 @@ public final class SubClass
 		return _exp;
 	}
 	
-	public int getSp()
+	public long getSp()
 	{
 		return _sp;
 	}
@@ -105,7 +105,7 @@ public final class SubClass
 		_exp = expValue;
 	}
 	
-	public void setSp(int spValue)
+	public void setSp(long spValue)
 	{
 		_sp = spValue;
 	}
