+++ L2J_Mobius_CT_2.6_HighFive/java/com/l2jmobius/gameserver/network/serverpackets/GMViewCharacterInfo.java
@@ -68,7 +68,7 @@ public class GMViewCharacterInfo implements IClientOutgoingPacket
 		packet.writeD((int) _activeChar.getCurrentHp());
 		packet.writeD(_activeChar.getMaxMp());
 		packet.writeD((int) _activeChar.getCurrentMp());
-		packet.writeD(_activeChar.getSp());
+		packet.writeD((int) _activeChar.getSp());
 		packet.writeD(_activeChar.getCurrentLoad());
 		packet.writeD(_activeChar.getMaxLoad());
 		packet.writeD(_activeChar.getPkKills());
